refac: modularized the schemes
