+++ src/docs/NoteSchemas.ts
@@ -1,34 +1,6 @@
 import { FastifySchema } from 'fastify';
 import { FastifyRouteSchemaDef } from 'fastify/types/schema';
-
-export const NoteSchema = {
-    $id: 'Note',
-    description: 'The `Note` object definition',
-    type: 'object',
-    required: ['title', 'description'],
-    properties: {
-        title: { type: 'string' },
-        description: { type: 'string' },
-    },
-};
-export const NoteResponseSchema = {
-    $id: 'NoteResponse',
-    description: 'Response object of `Note` resource',
-    type: 'object',
-    required: ['id', 'title', 'description'],
-    properties: {
-        id: { type: 'string' },
-        title: { type: 'string' },
-        description: { type: 'string' },
-    },
-};
-
-export const NoteUpdateSchema = {
-    $id: 'NoteUpdate',
-    description: 'Response object of `Note` resource',
-    type: 'object',
-    properties: NoteSchema.properties,
-};
+import { NoteResponseSchema, NoteSchema, NoteUpdateSchema } from './schemas';
 
 const getNoteByIdSchema: FastifySchema = {
     description:
