+++ src/docs/index.ts
@@ -1,14 +1,8 @@
 import { FastifyInstance } from 'fastify';
-import {
-    NoteResponseSchema,
-    NoteSchema,
-    NoteUpdateSchema,
-} from './NoteSchemas';
+import * as schemas from './schemas';
 
 export * as NoteRouteSchema from './NoteSchemas';
 
 export const defineSchemas = (app: FastifyInstance): void => {
-    app.addSchema(NoteSchema);
-    app.addSchema(NoteResponseSchema);
-    app.addSchema(NoteUpdateSchema);
+    Object.values(schemas).forEach(schema => app.addSchema(schema));
 };
